issues-10-tabbar: Added tab bar.
