+++ IceAndFire/app/App.js
@@ -8,11 +8,10 @@ import React, { Component } from 'react';
 
 import Main from './components/Main'
 
-export default class App extends Component<{}> {
+export default class App extends Component {
   render() {
     return (
       <Main />
     );
   }
 }
-
