+++ IceAndFire/app/components/Main.js
@@ -1,41 +1,24 @@
 import React, { Component } from 'react';
-import {StyleSheet} from 'react-native';
-import ContainerList from './ListComponent';
+import HouseScreen from "./HouseScreen";
+import BookScreen from "./BookScreen";
+import { StackNavigator, TabNavigator } from 'react-navigation';
 
 export default class Main extends Component {
-
-    state = {
-        headerTextAlign: 'center', //`left`,
-        headerText: 'Header Goes here',
-        headerTextDecoration: 'none', //'underline',
-        headerTextFontSize: 18 //20
-    }
-
     render() {
         return (
-            <ContainerList 
-                containerStyle={styles.container} 
-                headerStyle = {[
-                    styles.header, 
-                    {textAlign: this.state.headerTextAlign}, 
-                    {textDecorationLine: this.state.headerTextDecoration},
-                    {fontSize: this.state.headerTextFontSize}
-                ]}
-                headerText =  {this.state.headerText}
-            />
+            <RootNavigator/>
         );
     }
 }
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        marginTop: 20
+const MainScreenNavigator = TabNavigator({
+    House: { screen: HouseScreen },
+    Book: { screen: BookScreen },
+});
+
+const RootNavigator = StackNavigator({
+    Main: {
+        screen: MainScreenNavigator,
     },
-    header: {
-        flex: 0,
-        marginTop: 8,
-        fontWeight: 'bold'
-    }
 });
