+++ IceAndFire/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "react": "16.2.0",
     "react-native": "0.52.0",
+    "react-navigation": "^1.0.0-beta.22",
     "react-native-elements": "^0.19.0",
     "react-native-vector-icons": "^4.5.0"
   },
